+++ changelog.md
@@ -3,6 +3,10 @@
 - 2023-08-02
   - Fixing drag amount when moving boxes.
   - Drawing connections more beautiful with cubic Bézier curves.
+  - Feature to add new answers to a node.
+  - Feature to change answer order.
+  - Feature to remove answers.
+  - Feature to remove nodes.
 - 2023-08-01
   - Added DOMHelpers and a basic editor interface.
 - 2023-07-31